requirements.txt
# 视频处理模块依赖
ffmpeg-python==0.2.0
opencv-python==4.9.0.80
librosa==0.10.2.post1
PyYAML==6.0.1
numpy==1.24.3
soundfile==0.12.1
pydub==0.25.1

# 音频分离依赖
spleeter==2.4.2
tensorflow==2.12.1
pandas==1.5.3



